Refactor: final 폴더 내용 react 폴더로 통합 및 final 폴더 삭제
