--- final/src/main.jsx
@@ -1,8 +1,0 @@
-import { createRoot } from 'react-dom/client'
-import App from './App.jsx'
-
-createRoot(document.getElementById('root')).render(
-
-    <App />
-
-)
