+++ react/src/App.jsx
@@ -1,4 +1,15 @@
 import { BrowserRouter, Route, Routes } from "react-router-dom";
+import FreeBoard from "../../react/src/pages/freeBoard";
+import FreeboardDetail from "../../react/src/pages/freeboardDetail";
+import FreeboardEdit from "../../react/src/pages/freeboardEdit";
+import FreeboardWrite from "../../react/src/pages/freeboardWrite";
+import Notice from "../../react/src/pages/notice";
+import NoticeDetail from "../../react/src/pages/noticeDetail";
+import Qna from "../../react/src/pages/qna";
+import QnaDetail from "../../react/src/pages/qnaDetail";
+import QnaEdit from "../../react/src/pages/qnaEdit";
+import QnaWrite from "../../react/src/pages/qnaWrite";
+import Test from "../../react/src/pages/test";
 import './App.css';
 import Layout from "./layouts/Layout";
 import CampingPage from "./pages/CampingPage";
@@ -9,8 +20,6 @@ import MountainPage from "./pages/MountainPage";
 import Mypage from "./pages/Mypage";
 import ShortsVideoPage from "./pages/ShortsVideoPage";
 import ThemeParkPage from "./pages/ThemeParkPage";
-import Bookmark from './pages/bookmark';
-import Search from './pages/search';
 
 function App() {
 
@@ -27,10 +36,29 @@ function App() {
         <Route path="/area" element={<GangwondoPage />} />
 
         <Route path="/mypage" element={<div className="mypage"><Mypage /></div>} />
-        <Route path="/" element={<div className="search-wrapper"><Search /></div>} />
-          <Route path="/search" element={<div className="search-wrapper"><Search /></div>} />
-          <Route path="/bookmark" element={<div className="bookmark-wrapper"><Bookmark /></div>} />
-          <Route path="/mypage" element={<div className="mypage"><Mypage /></div>} />
+
+        <Route path="/notice" element={<Notice />} />
+
+        <Route path="/noticeDetail" element={<NoticeDetail />} />
+
+        <Route path="/freeboard" element={<FreeBoard />} />
+
+        <Route path="/test" element={<Test />} />
+
+        <Route path="/freeboardDetail" element={<FreeboardDetail />} />
+
+        <Route path="/freeboardWrite" element={<FreeboardWrite />} />
+
+        <Route path="/freeboardEdit" element={<FreeboardEdit />} />
+
+        <Route path="/qna" element={<Qna />} />
+
+        <Route path="/qnaDetail" element={<QnaDetail />} />
+
+        <Route path="/qnaWrite" element={<QnaWrite />} />
+
+        <Route path="/qnaEdit" element={<QnaEdit />} />
+
       </Routes>
     </BrowserRouter>
   )
