+++ react/src/pages/freeboardDetail.jsx
