+++ react/src/pages/freeboardDetail.module.css
