+++ react/src/pages/noticeDetail.jsx
