+++ react/src/pages/qna.jsx
@@ -1,4 +1,4 @@
-import Pagination from "../common/Pagination";
+import Pagination from "../common/pagination";
 
 import qnaStyle from "./qna.module.css";
