+++ react/src/pages/qnaDetail.jsx
