+++ react/src/pages/qnaEdit.module.css
@@ -99,9 +99,9 @@
   }
   
   /* 이미지 첨부(img) 영역 */
-  .img {
-    /* 특별한 스타일이 현재는 필요 없어 보이지만, 필요시 추가 */
-  }
+  /* .img {
+    특별한 스타일이 현재는 필요 없어 보이지만, 필요시 추가
+  } */
   
   .imagePreviewContainer {
     display: flex;
