+++ react/src/pages/qnaWrite.jsx
